Change token retry policy to use correct repo.
