+++ lib/core/network/token_interceptor.dart
@@ -68,31 +68,3 @@ class TokenInterceptor implements InterceptorContract {
   @override
   Future<bool> shouldInterceptResponse() => Future.value(true);
 }
-
-/// This is the retry policy that will be used by the [InterceptedClient]
-/// to retry requests that failed due to an expired token.
-class ExpiredTokenRetryPolicy extends RetryPolicy {
-  @override
-  int maxRetryAttempts = 2;
-
-  @override
-  Future<bool> shouldAttemptRetryOnResponse(BaseResponse response) async {
-    ///This is where we need to update our token on 401 response
-    if (response.statusCode == 401) {
-      await _refreshToken();
-      return true;
-    }
-    return false;
-  }
-
-  /// This method will be called
-  /// when a request fails and the [shouldAttemptRetryOnResponse]
-  /// Handle the [SocketException] when there is no internet connection
-  Future<void> _refreshToken() async {
-    try {
-      await getIt<AuthRepository>().refreshToken();
-    } catch (e) {
-      log('No internet connection');
-    }
-  }
-}
